+++ crates/miden-testing/src/kernel_tests/tx/test_tx.rs
@@ -166,9 +166,10 @@ async fn consuming_note_created_in_future_block_fails() -> anyhow::Result<()> {
     // Attempt to execute a transaction against reference block 1 with the note created in block 11
     // - which should fail.
     let tx_context = mock_chain.build_tx_context(account.id(), &[], &[])?.build()?;
-    let source_manager = tx_context.source_manager();
 
-    let tx_executor = TransactionExecutor::<'_, '_, _, UnreachableAuth>::new(&tx_context, None);
+    let tx_executor = TransactionExecutor::<'_, '_, _, UnreachableAuth>::new(&tx_context)
+        .with_source_manager(tx_context.source_manager());
+
     // Try to execute with block_ref==1
     let error = tx_executor
         .execute_transaction(
@@ -176,7 +177,6 @@ async fn consuming_note_created_in_future_block_fails() -> anyhow::Result<()> {
             BlockNumber::from(1),
             InputNotes::new(vec![input_note]).unwrap(),
             TransactionArgs::default(),
-            source_manager,
         )
         .await;
 
@@ -975,6 +975,7 @@ async fn advice_inputs_from_transaction_witness_are_sufficient_to_reexecute_tran
             acct_procedure_index_map,
             None,
             tx_inputs.block_header().fee_parameters(),
+            Arc::new(DefaultSourceManager::default()),
         )
     };
     let advice_inputs = advice_inputs.into_advice_inputs();
@@ -1400,7 +1401,6 @@ async fn execute_tx_view_script() -> anyhow::Result<()> {
     ";
 
     let source = NamedSource::new("test::module_1", test_module_source);
-    let source_manager = Arc::new(DefaultSourceManager::default());
     let assembler = TransactionKernel::assembler();
 
     let library = assembler.assemble_library([source]).unwrap();
@@ -1426,17 +1426,10 @@ async fn execute_tx_view_script() -> anyhow::Result<()> {
     let block_ref = tx_context.tx_inputs().block_header().block_num();
     let advice_inputs = tx_context.tx_args().advice_inputs().clone();
 
-    let executor = TransactionExecutor::<'_, '_, _, UnreachableAuth>::new(&tx_context, None);
+    let executor = TransactionExecutor::<'_, '_, _, UnreachableAuth>::new(&tx_context);
 
     let stack_outputs = executor
-        .execute_tx_view_script(
-            account_id,
-            block_ref,
-            tx_script,
-            advice_inputs,
-            Vec::default(),
-            source_manager,
-        )
+        .execute_tx_view_script(account_id, block_ref, tx_script, advice_inputs, Vec::default())
         .await?;
 
     assert_eq!(stack_outputs[..3], [Felt::new(7), Felt::new(2), ONE]);