+++ CHANGELOG.md
@@ -74,6 +74,7 @@
 - [BREAKING] Update `account::get_storage_commitment` procedure to `account::compute_storage_commitment`([#1763](https://github.com/0xMiden/miden-base/pull/1763)).
 - Implement caching for the account storage commitment (([#1763](https://github.com/0xMiden/miden-base/pull/1763))).
 - [BREAKING] Merge the current and initial account code commitment procedures into one ([#1776](https://github.com/0xMiden/miden-base/pull/1776)).
+- [BREAKING] Made source manager an instance variable of the `TransactionExecutor` ([#1788](https://github.com/0xMiden/miden-base/pull/1788)).
 
 ## 0.10.1 (2025-08-02)